requirements.txt
# ScholarSync Dependencies
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1

# LLM & Agents
groq==0.13.1
langgraph==0.2.62
langchain-core==0.3.29

# RAG & Embeddings
llama-index-core==0.12.12
chromadb==0.5.23
sentence-transformers==3.3.1

# Graph Database
neo4j==5.27.0

# Document Processing
PyMuPDF==1.25.2

# Utilities
python-dotenv==1.0.1
